requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
PyPDF2==3.0.1
python-docx==1.1.2
nltk==3.9.1
scikit-learn==1.6.1
numpy==2.2.3
sentence-transformers==3.4.1
requests==2.32.3
beautifulsoup4==4.12.3
reportlab==4.2.5
python-pptx==1.0.2
Jinja2==3.1.5
gunicorn==23.0.0
